+++ Policyfile.rb
@@ -10,8 +10,7 @@ name 'app_ror'
 default_source :supermarket
 
 # run_list: chef-client will run these recipes in the order specified.
-run_list 'app_ror::default'
-named_run_list :test, 'test::default'
+run_list 'test::default'
 
 # Specify a custom source for a single cookbook:
 cookbook 'app_ror', path: '.'