+++ recipes/default.rb
@@ -1,5 +1,0 @@
-#
-# Cookbook:: app_ror
-# Recipe:: default
-#
-# Copyright:: 2021, Earth U