+++ resources/logrotate.rb
@@ -2,7 +2,19 @@
 # Cookbook:: app_ror
 # Resource:: logrotate
 #
-# Copyright:: 2021, Earth U
+# Copyright:: 2022, Earth U
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 cb = 'app_ror'
 unified_mode true
@@ -16,42 +28,31 @@ property :path, [String, Array],
          default: []
 
 property :config, Array,
-         description: 'Logrotate settings for this config'
+         description: 'Logrotate settings for this config',
+         default: [
+           'weekly',
+           'missingok',
+           'rotate 12',
+           'compress',
+           'delaycompress',
+           'notifempty',
+           'copytruncate',
+         ]
 
 property :logrotate_dir, String,
-         description: 'Default location of logrotate files',
+         description: 'Location of logrotate config files',
          default: '/etc/logrotate.d'
 
-action_class do
-  def prop_path
-    [new_resource.path].flatten
-  end
-
-  def prop_config
-    if property_is_set?(:config)
-      new_resource.config
-    else
-      [
-        'weekly',
-        'missingok',
-        'rotate\ 12',
-        'compress',
-        'delaycompress',
-        'notifempty',
-        'copytruncate',
-      ]
-    end
-  end
-end
-
 action :create do
-  unless prop_path.empty?
+  path = [new_resource.path].flatten
+
+  unless path.empty?
     template "#{new_resource.logrotate_dir}/#{new_resource.conf_name}" do
       cookbook cb
       source   'logrotate.conf.erb'
       variables(
-        path:   prop_path,
-        config: prop_config
+        path:   path,
+        config: new_resource.config
       )
     end
   end