+++ test/cookbooks/test/attributes/default.rb
@@ -1,9 +1,8 @@
 #
-# Author:: Earth U (<iskitingbords@gmail.com>)
-# Cookbook Name:: test
-# Recipe:: ruby
+# Cookbook:: test
+# Attribute:: default
 #
-# Copyright (C) 2018, Earth U
+# Copyright:: 2022, Earth U
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,11 +15,16 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
-app_ror_ruby node['test']['ruby_ver'] do
-  user     node['test']['user']
-  prefix   node['test']['ruby_prefix']
-  gem_path node['test']['ruby_gem_path']
-  gems     node['test']['ruby_gems']
-end
+default['test']['base_dir'] = '/var/src/myapp'
+default['test']['sub_dirs'] = %w(
+  shared
+  shared/config
+  shared/tmp
+  shared/log
+)
+
+default['test']['ruby_ver'] = '2.6.10'
+default['test']['ruby_env'] = { 'TEST_RUBY_VAR' => 'foobar' }
+
+default['test']['sidekiq_dependencies'] = 'redis-server.service'