+++ test/cookbooks/test/recipes/default.rb
@@ -1,9 +1,8 @@
 #
-# Author:: Earth U (<iskitingbords@gmail.com>)
-# Cookbook Name:: test
+# Cookbook:: test
 # Recipe:: default
 #
-# Copyright (C) 2018, Earth U
+# Copyright:: 2022, Earth U
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,12 +15,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
 # include_recipe 'test::swap'
-# include_recipe 'test::base_dirs'
-# include_recipe 'test::logrotate'
-# include_recipe 'test::ruby'
+app_ror_base_dirs node['test']['base_dir'] do
+  sub_dirs node['test']['sub_dirs']
+end
+
+include_recipe 'test::ruby'
+include_recipe 'test::manage_puma'
+include_recipe 'test::manage_sidekiq'
+include_recipe 'test::logrotate'
 # include_recipe 'test::solr'
-# include_recipe 'test::manage_puma'
-# include_recipe 'test::manage_sidekiq'