+++ test/cookbooks/test/recipes/manage_puma.rb
@@ -1,9 +1,8 @@
 #
-# Author:: Earth U (<iskitingbords@gmail.com>)
-# Cookbook Name:: test
+# Cookbook:: test
 # Recipe:: manage_puma
 #
-# Copyright (C) 2018, Earth U
+# Copyright:: 2022, Earth U
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,8 +15,5 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
-app_ror_manage_puma "#{node['test']['base_dir']}/current" do
-  user node['test']['user']
-end
+app_ror_manage_puma "#{node['test']['base_dir']}/current"